% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_number.R
\name{aggregate_number}
\alias{aggregate_number}
\title{Aggregate replicate simulations}
\usage{
aggregate_number(
  d,
  theta = 0.9,
  xname = "hsymmetry",
  yname = "ecosel",
  zname = "summary",
  colname = NULL,
  keep = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{theta}{The threshold to pass}

\item{xname}{Column name of the first grouping parameter}

\item{yname}{Column name of the second grouping parameter (set to NULL if only one grouping parameter)}

\item{zname}{Column name of the statistic to aggregate}

\item{colname}{Optional name of the aggregated statistic column}

\item{keep}{Optional names of columns to keep in the aggregated data frame}
}
\description{
Aggregate a statistic over replicates by counting how many have values above a threshold, within values of one or two parameters.
}
