% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_quantile.R
\name{aggregate_quantile}
\alias{aggregate_quantile}
\title{Aggregate replicate simulations}
\usage{
aggregate_quantile(
  d,
  prob = 0.95,
  xname = "hsymmetry",
  yname = "ecosel",
  zname = "summary",
  colname = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{prob}{Which quantile? (defaults to 0.95)}

\item{xname}{Column name of the first grouping parameter}

\item{yname}{Column name of the second grouping parameter (set to NULL if only one grouping parameter)}

\item{zname}{Column name of the statistic to aggregate}

\item{colname}{Optional name of the aggregated statistic column}
}
\description{
Aggregate a statistic to a quantile over replicates within values of one or two parameters.
}
