% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect simulation data}
\usage{
collect(dir = ".", xfile = "EI.dat", yfile = "SI.dat", zfile = "RI.dat")
}
\arguments{
\item{dir}{Where the simulation folders are}

\item{xfile}{Filename for the first dimension}

\item{yfile}{Filename for the second dimension}

\item{zfile}{Filename for the third dimension}
}
\description{
Gathers simulation data through time from multiple simulations.
}
