% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_params.R
\name{collect_params}
\alias{collect_params}
\title{Collect simulation parameters}
\usage{
collect_params(parnames, dir = ".", parfile = "paramlog.txt")
}
\arguments{
\item{parnames}{What parameters to extract}

\item{dir}{Where the simulation folders are}

\item{parfile}{Name of the parameter log file}
}
\description{
Gathers simulation parameters for multiple simulations.
}
