% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine datasets}
\usage{
combine_data(d, levs = NULL)
}
\arguments{
\item{d}{A list of data frames to combine. Numbers of columns should match.}

\item{levs}{Optional list representing columns to add, where each element contains the levels to assign to each data frame.}
}
\description{
Combine multiple data frames and add extra columns to identify them.
}
