% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrep.R
\name{mrep}
\alias{mrep}
\title{Repeat multiple elements multiple times}
\usage{
mrep(vecx, vecy)
}
\arguments{
\item{vecx}{Vector of elements}

\item{vecy}{Vector of times}
}
\description{
Repeat multiple elements multiple times
}
