% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{myfun}
\alias{myfun}
\title{Summarize simulations}
\usage{
myfun(x, xnames = NULL, tname = "t", from = NULL, to = NULL)
}
\arguments{
\item{tname}{Optional time column}

\item{d}{Data frame with coordinates per timepoint per simulation}

\item{colvar}{What variable to color according to}

\item{labs}{Labels for each axis}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}

\item{splitvar}{Facet splitting variable}

\item{splitvar2}{Second fact splitting variable}
}
\description{
Summarize simulations through time by a single value.
}
