% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cube.R
\name{plot_cube}
\alias{plot_cube}
\title{Plot cube}
\usage{
plot_cube(d, labs, xname = "x", yname = "y", zname = "z", ...)
}
\arguments{
\item{d}{Data frame with three coordinates per timepoint per simulation}

\item{labs}{Labels for each axis}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}

\item{zname}{Column name of the z-axis}

\item{...}{Parameters to be passed to text3D}
}
\description{
Plots simulations in a speciation cube.
}
