% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap of a statistic across parameter space}
\usage{
plot_heatmap(
  d,
  labs = NULL,
  xname = "hsymmetry",
  yname = "ecosel",
  zname = "z",
  tname = "t",
  summary = "value",
  tval = NULL,
  trange = NULL,
  theta = 0.9,
  colname = NULL,
  aggregate = "average",
  prob = 0.95,
  threshold = 0.9,
  colors = NULL,
  collab = NULL,
  splitvar = NULL,
  splitvar2 = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{labs}{Labels for each axis}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}

\item{zname}{Column name of the variable to summarize}

\item{tname}{Time column}

\item{summary}{Method used to summarize simulations. Either of "value" (picks a value at a given time point), "average" (between two time points) or "threshold" (the time at which a threshold is passed, assuming monotonous increase).}

\item{tval}{Timepoint to pick a value at (defaults to last generation). Used only if summary method is "value".}

\item{trange}{Timepoints between which to average (defaults to the entire range). Used only if summary method is "average".}

\item{theta}{Threshold whose passing time (assuming monotonous increase) is recorded. Used only if summary method is "threshold".}

\item{colname}{Optional name of the summary column}

\item{aggregate}{Method used to aggregate replicate simulations. Either of "average", "variance", "quantile" or "number" (the number of replicates above a certain threshold of the summary statistic).}

\item{prob}{The quantile used to aggregate replicates. Used only if the aggregation method is "quantile".}

\item{threshold}{The threshold used if the aggregation method is "number".}

\item{colors}{Optional low and high ends of the color gradient}

\item{collab}{Optional color legend title}

\item{splitvar}{Facet splitting variable}

\item{splitvar2}{Second facet splitting variable}
}
\description{
Compute a summary statistic for each simulation, then aggregates replicate simutions per combination of parameters and plot the result.la
}
