% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{read}
\alias{read}
\title{Collect simulation data}
\usage{
read(filename)
}
\arguments{
\item{dir}{Where the simulation folders are}

\item{xfile}{Filename for the first dimension}

\item{yfile}{Filename for the second dimension}

\item{zfile}{Filename for the third dimension}

\item{parnames}{Optional names of the parameters to read and append}

\item{parfile}{Name of the file where to read parameter values. Used only if parameter names are provided.}
}
\description{
Gathers simulation data through time from multiple simulations.
}
\note{
The function will look for the "sim_" pattern to identify the folders containing simulation data. Make sure that the full path of the directory where the simulation folders are does not contain another "sim_" string, otherwise all files in the directory will be interpreted as simulation folders, and errors may occur.
}
