% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_params}
\alias{read_params}
\title{Read simulation parameters}
\usage{
read_params(parnames, parfile = "paramlog.txt", dir = ".")
}
\arguments{
\item{parnames}{The names of the parameters to read}

\item{parfile}{The name of the parameter file}

\item{dir}{Where the parameter file is located. Defaults to the current working directory.}
}
\description{
Reads the values of parameters of interest for a given simulation.
}
