% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_average.R
\name{summ_average}
\alias{summ_average}
\title{Summarize simulations}
\usage{
summ_average(d, sname = "x", tname = "t", trange = NULL, colname = NULL)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{sname}{Column name of the statistic to summarize}

\item{tname}{Time column}

\item{trange}{The two values between which to average (defaults to the entire range)}

\item{colname}{Optional name of the summary column}
}
\description{
Summarize simulations down to the average value of a statistic between two time points.
}
