% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_threshold.R
\name{summ_threshold}
\alias{summ_threshold}
\title{Summarize simulations}
\usage{
summ_threshold(d, sname = "x", tname = "t", theta = 0.9, colname = NULL)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{sname}{Column name of the statistic to summarize}

\item{tname}{Time column}

\item{theta}{The threshold that needs to be passed}

\item{colname}{Optional name of the summary column}
}
\description{
Summarize simulations down to the time point at which a statistic passes a certain threhold (assuming monotonous increase).
}
