% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_value.R
\name{summ_value}
\alias{summ_value}
\title{Summarize simulations}
\usage{
summ_value(
  d,
  sname = "x",
  tname = "t",
  tval = NULL,
  colname = NULL,
  keep = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{sname}{Column name of the statistic to summarize}

\item{tname}{Time column}

\item{tval}{Timepoint to pick (defaults to last generation)}

\item{colname}{Optional name of the summary column}

\item{keep}{Optional names of columns to keep in the summary data frame}
}
\description{
Summarize simulations down to the value of a statistic picked at a given time.
}
