% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect simulation data}
\usage{
collect(
  dir = ".",
  files = c("EI.dat", "SI.dat", "RI.dat"),
  parnames = c("hsymmetry", "ecosel"),
  parfile = "paramlog.txt"
)
}
\arguments{
\item{dir}{Where the simulation folders are}

\item{files}{Filenames of the data to extract}

\item{parnames}{Optional names of the parameters to read and append}

\item{parfile}{Name of the file where to read parameter values. Used only if parameter names are provided.}
}
\description{
Gathers simulation data through time from multiple simulations.
}
\note{
It is important to not have the string "sim_" in the path leading to where the simulation folders are. This is because the function to collect data looks for simulation folders using the pattern sim_ in the full path of those folders. If sim_ appears somewhere else in the path than in the name of the simulation folders, all files in the directory containing the simulation folders will be interpreted as simulation folders, even if they are not, e.g. python scripts.
}
