% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot speciation heatmap}
\usage{
plot_heatmap(
  d,
  zname,
  labs = NULL,
  xname = "x",
  yname = "y",
  splitvar = NULL,
  splitvar2 = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{zname}{Column name of the heat variable}

\item{labs}{Labels for each axis}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}

\item{splitvar}{Facet splitting variable}

\item{splitvar2}{Second facet splitting variable}
}
\description{
Summarizes simulation metrics in a heatmap.
}
