% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plane.R
\name{plot_plane}
\alias{plot_plane}
\title{Plot speciation phase plane}
\usage{
plot_plane(
  d,
  xname = "x",
  yname = "y",
  tname = NULL,
  labs = NULL,
  colvar = NULL,
  collab = NULL,
  splitvar = NULL,
  splitvar2 = NULL
)
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}

\item{tname}{Optional time column}

\item{labs}{Labels for each axis}

\item{colvar}{What variable to color according to}

\item{collab}{Optional color legend title}

\item{splitvar}{Facet splitting variable}

\item{splitvar2}{Second facet splitting variable}
}
\description{
Plots simulations in a phase plane. If tname is provided, plots y against time.
}
