% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plane.R
\name{plot_plane}
\alias{plot_plane}
\title{Plot speciation phase plane}
\usage{
plot_plane(d, xname = "x", yname = "y")
}
\arguments{
\item{d}{Data frame with coordinates per timepoint per simulation}

\item{xname}{Column name of the x-axis}

\item{yname}{Column name of the y-axis}
}
\description{
Plots simulations in a phase plane.
}
